rag-backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Vector Database
qdrant-client==1.7.3

# AI/ML Services
cohere==4.47.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
tiktoken==0.5.2
langdetect==1.0.9
python-frontmatter==1.0.1

# HTTP Client
httpx==0.26.0

# Logging
python-json-logger==2.0.7

# Testing (optional for Phase 6)
pytest==7.4.4
pytest-asyncio==0.23.3
